# Faster iteration for an IdOffsetRange