--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -221,6 +221,7 @@ end
             @test OrdinalRange{BigInt,BigInt}(r2) === r2
         end
     end
+<<<<<<< HEAD
 
     @testset "Bool IdOffsetRange (issue #223)" begin
         for b1 in [false, true], b2 in [false, true]
@@ -344,6 +345,25 @@ end
             @test_throws BoundsError r[true:true:false]
             @test_throws BoundsError r[false:true:false]
         end
+=======
+    @testset "iteration" begin
+        # parent has Base.OneTo axes
+        A = ones(4:10)
+        ax = axes(A, 1)
+        ind, st = iterate(ax)
+        @test A[ind] == A[4]
+        ind, st = iterate(ax, st)
+        @test A[ind] == A[5]
+
+        # parent doesn't have Base.OneTo axes
+        B = @view A[:]
+        C = OffsetArray(B, 0)
+        ax = axes(C, 1)
+        ind, st = iterate(ax)
+        @test C[ind] == C[4]
+        ind, st = iterate(ax, st)
+        @test C[ind] == C[5]
+>>>>>>> df172e1... specialize for Base.OneTo
     end
 end