Fixes #214 

This was not an issue with bounds-checking, but with iterating over an `IdOffsetRange`. Instead we may convert it to a `UnitRange` while iterating. 

After this:

```julia
julia> x = zeros(Int, 10_000_000);

julia> y = OffsetVector(x, 0);

julia> fill1d(x) = for i in axes(x,1); x[i] = i; end
fill1d (generic function with 1 method)

julia> @btime fill1d($x);
  11.666 ms (0 allocations: 0 bytes)

julia> @btime fill1d($y);
  11.884 ms (0 allocations: 0 bytes)
```