@timholy 2021-03-01
It seems that there's a failure to automatically elide the bounds-checking. I'd rather figure out why.

One concern with solutions like this is that it potentially changes behavior. What if you can't coerce to a `UnitRange`? What if you define an `AbstractRange` type with side effects?

```julia
tim@diva:/tmp$ juliam --startup-file=no -q
julia> struct MyUR <: AbstractUnitRange{Int} end

julia> Base.iterate(::MyUR) = (println("starting"); return (1, false))

julia> Base.iterate(::MyUR, isdone::Bool) = (println("continuing"); return isdone ? nothing : (2, true))

julia> Base.first(::MyUR) = 1

julia> Base.last(::MyUR) = 2

julia> r = MyUR();

julia> for i in r
           println(i)
       end
starting
1
continuing
2
continuing

julia> UnitRange(r)
1:2
```
so `UnitRange(r)` destroys any special behavior of the parent iterator.