@jishnub 2021-03-02
I think this is more about iteration than bounds checking. The way I look at this is: `IdOffsetRange` is an internal type to `OffsetArrays`, and the package is free to define how to iterate over it to suit its goals. A wrapper type is not expected to have a one-to-one correspondence with the parent. If one desires the iteration behavior of `MyUR`, they may iterate over the parent or construct an appropriate type.

Regarding the possibility of coercion to a `UnitRange`: this requires that the parent defines its first and last values, which should be common for `AbstractUnitRange`s. In any case this is not strictly necessary, we may use the fact that `IdOffsetRange` has a unit step to define the iteration.