@jishnub 2021-03-04
For reasons that I don't fully understand, bounds checking and populating a 2d array is faster if we coerce the axes to `UnitRange`s before iterating over them.

On master:
```julia
julia> x2d = zeros(Int, 4000, 4000);

julia> y2d = OffsetArray(x2d, 0, 0);

julia> fill2d(x) = for j in axes(x,2); for i in axes(x,1); x[i,j] = i + j; end; end
fill2d (generic function with 1 method)

julia> @btime fill2d($x2d);
  17.113 ms (0 allocations: 0 bytes)

julia> @btime fill2d($y2d);
  31.932 ms (0 allocations: 0 bytes)
```

Using e8e184c:
```julia
julia> @btime fill2d($y2d);
  19.826 ms (0 allocations: 0 bytes)
```

After c3e04b5
```julia
julia> @btime fill2d($y2d);
  17.588 ms (0 allocations: 0 bytes)
```

@timholy Could you investigate this performance issue when you have time? This is a significant improvement that's possible.