@johnnychen94 2021-03-18
Surprisingly, by changing `e8e184c` to

```diff
- Base.iterate(r::IdOffsetRange) = isempty(r) ? nothing : (first(r), first(r))
+ @inline function Base.iterate(r::IdOffsetRange)
+   if isempty(r)
+       return nothing
+   else
+       next = first(r)
+       return (next, next)
+   end
+ end
```

I get

```julia
julia> @btime fill2d($x2d);
  7.454 ms (0 allocations: 0 bytes)

julia> @btime fill2d($y2d);
  7.685 ms (0 allocations: 0 bytes)
  10.894 ms (0 allocations: 0 bytes) # Using e8e184c
```

I guess the still existing 0.3ms gap lies in `last(r)`, where the extra computation seems unavoidable if we still want to keep generality.

> One concern with solutions like this is that it potentially changes behavior. What if you can't coerce to a UnitRange? What if you define an AbstractRange type with side effects?

I feel it be okay to specialize the method for common use cases that can be converted to a UnitRange.