@jishnub 2021-03-18
The common parent range is perhaps `Base.OneTo` and not `UnitRange`. In essence this is equivalent to an `UnitRange`, so such an approach makes sense. However it'll be better if this can be solved generally.