@jishnub 2021-04-11
I've added a specialized method for `Base.OneTo` but this is not really ideal