@jishnub 2021-07-05
With the iteration defined as it is now

```julia
julia> dim = 1000
1000

julia> x2d = Array{Float64}(undef, 2dim, 2dim);

julia> fill2d(x) = for j in axes(x,2); for i in axes(x,1); x[i,j] = i + j; end; end
fill2d (generic function with 1 method)

julia> @btime fill2d($x2d);
  5.247 ms (0 allocations: 0 bytes)

julia> @btime fill2d($(OffsetArray(x2d)));
  4.748 ms (0 allocations: 0 bytes)
```

So wrapping an `OffsetArray` actually makes the function faster. This is unusual and other functions in the benchmark don't behave this way, however this is interesting indeed.