@jishnub 2021-07-05
Good to merge this? There's an evident performance boost for `OneTo` and no change to the general case